+++ tests/test_schema_models.py
@@ -1,5 +1,5 @@
 def test_type_ref_empty():
-    from qlient.schema.models import TypeRef
+    from qlient.schema.types import TypeRef
     type_ref = TypeRef()
     assert type_ref.kind is None
     assert type_ref.name is None
@@ -8,7 +8,7 @@ def test_type_ref_empty():
 
 
 def test_type_ref_values():
-    from qlient.schema.models import TypeRef
+    from qlient.schema.types import TypeRef
     type_ref = TypeRef(kind="NON_NULL", name="String")
     assert type_ref.kind == "NON_NULL"
     assert type_ref.name == "String"
@@ -18,7 +18,7 @@ def test_type_ref_values():
 
 
 def test_type_ref_nested():
-    from qlient.schema.models import TypeRef
+    from qlient.schema.types import TypeRef
     type_ref = TypeRef(kind="NON_NULL", name=None, ofType=TypeRef(kind="LIST", name="String"))
     assert type_ref.kind == "NON_NULL"
     assert type_ref.name is None
@@ -28,7 +28,7 @@ def test_type_ref_nested():
 
 
 def test_type_ref_parse():
-    from qlient.schema.models import TypeRef
+    from qlient.schema.types import TypeRef
     type_ref = TypeRef.parse({"kind": "NON_NULL", "name": "String"})
     assert type_ref.kind == "NON_NULL"
     assert type_ref.name == "String"
@@ -38,7 +38,7 @@ def test_type_ref_parse():
 
 
 def test_input_empty():
-    from qlient.schema.models import Input
+    from qlient.schema.types import Input
     input_type = Input()
     assert input_type.name is None
     assert input_type.description is None
@@ -47,6 +47,6 @@ def test_input_empty():
 
 
 def test_input_values():
-    from qlient.schema.models import Input
+    from qlient.schema.types import Input
     input_type = Input(name="first_name")
     assert input_type.name == "first_name"
